refactor(lifecycle): pause and unpause move to their own file
commands.rs passed the 500-line limit with unpause_paused. No behaviour
changes.
